refactor: extract window creation into windows/ modules
Extract four window creation functions from main/index.ts into
separate modules in main/windows/, each receiving AppState as a
dependency:

- app-window.ts: createAppWindow(show, state, debug)
- editor-window.ts: openEditorWindow(state)
- install-window.ts: openInstallWindow(packId, state)
- debug-window.ts: createDebugWindow(state, debug, debugConfigFile)

Also extract DebugState interface into debug-state.ts for shared use.

main/index.ts now imports and delegates to these modules, reducing
its responsibilities to orchestration, IPC handlers, and state wiring.
