+++ src/main/windows/debug-state.ts
@@ -1,0 +1,7 @@
+export interface DebugState {
+  enabled: boolean;
+  identifier: string | undefined;
+  remoteUrl: string;
+  enable(): Promise<void>;
+  disable(): void;
+}
